Associate tool versions and external build output/dependencies with projects (#56, #57)
- crates/core/src/toolchain_declarations.rs (#56): read-only parsers
  for .tool-versions/mise.toml, .python-version, .ruby-version,
  .nvmrc/.node-version, rust-toolchain(.toml), plus rustup's global
  default_toolchain. Matches a declared version_spec to measured
  installations with manager semantics: an alias/range (lts/*, a bare
  3.12) stays an explicit UnresolvedRange unless exactly one
  installation uniquely matches it -- never a guessed closest version.
  Global-default scope is kept structurally distinct from per-project
  scope so many consuming projects and one default never collapse
  into one "consumer". Conflicting declarations within one project
  (e.g. .python-version vs .tool-versions naming different versions)
  are named explicitly.
- crates/core/src/external_associations.rs (#57): Xcode DerivedData
  info.plist WorkspacePath extraction (via the bounded, read-only,
  output-only `plutil -convert xml1 -o -`, never a bespoke binary-
  plist reader) joined against known project roots -- never inferred
  from the DerivedData folder's own basename+hash. Lockfile identity
  parsers for Cargo.lock, package-lock.json, pnpm-lock.yaml (narrow
  packages: scan, matching the existing .condarc precedent for
  non-general YAML), go.sum and gradle.lockfile, joined to projects
  via exact name+version match (never a similar-name guess). An
  unparseable lockfile is a named Unavailable gap, never a silent
  empty list. docker_join_evidence normalizes the existing Docker
  join logic's outcome into the same Consumer-evidence contract
  without re-deriving it.

Tests cover two projects sharing one Cargo.lock entry, a similar name
at a different version (not a match), a moved/missing Xcode project
(never guessed from basename), invalid lockfile metadata, nvm alias
resolution never guessing from installed versions present, and
global-default vs. per-project declaration scope.

Known gap (named, not silent): pom.xml dependency parsing is not
implemented (no XML parser dependency in this workspace); Maven
association still relies on recovery.rs's _remote.repositories marker
rather than a parsed dependency list.

+++ crates/core/src/activity.rs
@@ -317,7 +317,7 @@ fn parse_rfc3339_secs(s: &str) -> Option<u64> {
     // Hinnant's algorithm), avoiding a chrono dependency for one parse.
     let y = if month <= 2 { year - 1 } else { year };
     let era = if y >= 0 { y } else { y - 399 } / 400;
-    let yoe = (y - era * 400) as i64;
+    let yoe = y - era * 400;
     let mp = (month + 9) % 12;
     let doy = (153 * mp + 2) / 5 + day - 1;
     let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy;
